fix(masonkit): italic font-style, div color on iOS text, preflight buttons and lists
- Android stored Typeface constants in the font-style byte (italic read back
  as oblique); TS and iOS accessed the byte field as Int32, clobbering the
  neighbouring font state bytes. All use the 0/1/2 byte encoding now, and
  oblique synthesizes a slant on Android.
- iOS: drop cached default attributes before rebuilding text so inherited
  color reaches bare text; cap text width growth at the widest wrapped line.
- Preflight resets Button UA chrome and list markers on both platforms.
- Android: inline background spans for any non-Button container; rebuild
  parent inline segments when a span's font face loads.

+++ packages/nativescript-masonkit/src-native/mason-ios/Mason/Mason/MasonStyle.swift
@@ -1983,7 +1983,7 @@ public class MasonStyle: NSObject {
   
   public var fontStyle: FontStyle {
     get {
-      switch(getInt32(StyleKeys.FONT_STYLE_TYPE)){
+      switch(getInt8(StyleKeys.FONT_STYLE_TYPE)){
       case 0:
         return .Normal
       case 1:
